rename package to work with pip better
