+++ .gitignore
@@ -6,3 +6,4 @@ datastore/*
 __pycache__
 .pytest_cache
 build
+dist
