+++ changedetectionio/__init__.py
@@ -283,7 +283,7 @@ def changedetection_app(config=None, datastore_o=None):
             return response
 
         else:
-            from backend import forms
+            from changedetectionio import forms
             form = forms.quickWatchForm(request.form)
 
             output = render_template("watch-overview.html",
@@ -349,7 +349,7 @@ def changedetection_app(config=None, datastore_o=None):
     def get_current_checksum_include_ignore_text(uuid):
 
         import hashlib
-        from backend import fetch_site_status
+        from changedetectionio import fetch_site_status
 
         # Get the most recent one
         newest_history_key = datastore.get_val(uuid, 'newest_history_key')
@@ -376,7 +376,7 @@ def changedetection_app(config=None, datastore_o=None):
     @app.route("/edit/<string:uuid>", methods=['GET', 'POST'])
     @login_required
     def edit_page(uuid):
-        from backend import forms
+        from changedetectionio import forms
         form = forms.watchForm(request.form)
 
         # More for testing, possible to return the first/only
@@ -477,8 +477,8 @@ def changedetection_app(config=None, datastore_o=None):
     @login_required
     def settings_page():
 
-        from backend import forms
-        from backend import content_fetcher
+        from changedetectionio import forms
+        from changedetectionio import content_fetcher
 
         form = forms.globalSettingsForm(request.form)
 
@@ -734,7 +734,7 @@ def changedetection_app(config=None, datastore_o=None):
     @app.route("/api/add", methods=['POST'])
     @login_required
     def api_watch_add():
-        from backend import forms
+        from changedetectionio import forms
         form = forms.quickWatchForm(request.form)
 
         if form.validate():
@@ -851,7 +851,7 @@ def notification_runner():
         else:
             # Process notifications
             try:
-                from backend import notification
+                from changedetectionio import notification
                 notification.process_notification(n_object, datastore)
 
             except Exception as e:
@@ -861,7 +861,7 @@ def notification_runner():
 
 # Thread runner to check every minute, look for new watches to feed into the Queue.
 def ticker_thread_check_time_launch_checks():
-    from backend import update_worker
+    from changedetectionio import update_worker
 
     # Spin up Workers.
     for _ in range(datastore.data['settings']['requests']['workers']):
