+++ changedetectionio/static/images/favicon.ico
