+++ changedetectionio/static/images/gradient-border.png
