+++ changedetectionio/static/js/settings.js
