+++ changedetectionio/static/js/tabs.js
