+++ changedetectionio/static/styles/.gitignore
