+++ changedetectionio/static/styles/pure-min.css
