+++ changedetectionio/static/styles/styles.css
