+++ changedetectionio/static/styles/styles.scss
