+++ changedetectionio/templates/_helpers.jinja
