+++ changedetectionio/templates/edit.html
